Made LicePopulation use the refactored GenoDistrib
