+++ slim/simulation/config.py
@@ -20,7 +20,11 @@ from slim.types.treatments import (
 )
 
 if TYPE_CHECKING:
-    from slim.simulation.lice_population import LifeStage, GenoDistribDict
+    from slim.simulation.lice_population import (
+        LifeStage,
+        GenoDistribDict,
+        geno_config_to_matrix,
+    )
 
 
 def to_dt(string_date) -> dt.datetime:
@@ -171,10 +175,7 @@ class Config(RuntimeConfig):
 
         # Experiment-specific genetic ratios
         self.min_ext_pressure = data["ext_pressure"]
-        self.initial_genetic_ratios: GenoDistribDict = {
-            tuple(sorted(key.split(","))): val
-            for key, val in data["genetic_ratios"].items()
-        }
+        self.initial_genetic_ratios = geno_config_to_matrix(data["genetic_ratios"])
         self.genetic_learning_rate: float = data["genetic_learning_rate"]
         self.monthly_cost: float = data["monthly_cost"]
         self.gain_per_kg: float = data["gain_per_kg"]
